Heatmap Test
Trying to move the code that needs to be imported into heatmap_script into that file.
